feat: ダミーの画像を設定
